requirements.txt
python-dotenv
psycopg2
transformers
torch
black
pgvector
PyPDF2
bitsandbytes
accelerate
